pyproject.toml
[project]
name = "shai-hulud-detector"
version = "1.0.0"
description = "Shai Hulud Detector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "PyGithub>=2.1.0",
]
